- Implemented some hitTest prototypes

+++ view/gui/graphic_element/draw_node/draw_basin.cpp
@@ -2,7 +2,7 @@
 
 #include "view/gui/graphic_element/draw_primitive/draw_circle.h"
 
-DrawBasin::DrawBasin(float x, float y) : DrawNode(x, y) {
+DrawBasin::DrawBasin(double x, double y) : DrawNode(x, y) {
   initialize();
 }
